+++ MachineLearning.Data/Source/ImageDataSource.cs
@@ -6,7 +6,7 @@ using SixLabors.ImageSharp.Processing;
 namespace MachineLearning.Data.Source;
 
 /// <summary>
-/// reads all pngs in <paramref name="directoryInfo"/> and returns them as <see cref="double[]"/>
+/// reads all pngs in <paramref name="directoryInfo"/> and returns them as <see cref="float[]"/>
 /// </summary>
 public sealed class ImageDataSource(DirectoryInfo directoryInfo)
 {
@@ -16,18 +16,18 @@ public sealed class ImageDataSource(DirectoryInfo directoryInfo)
                 file.NameWithoutExtension().Parse<int>()
             )).ToArray();
 
-    public static float[] GetGrayscaleImageArray(FileInfo imageFile)
+    public static double[] GetGrayscaleImageArray(FileInfo imageFile)
     {
         using Image<Rgba32> image = Image.Load<Rgba32>(imageFile.FullName);
         image.Mutate(x => x.Grayscale());
 
-        var grayscaleValues = new float[image.Width * image.Height];
+        var grayscaleValues = new double[image.Width * image.Height];
 
         for(int y = 0; y < image.Height; y++)
         {
             for(int x = 0; x < image.Width; x++)
             {
-                grayscaleValues[y * image.Width + x] = image[x, y].R / 255.0f;
+                grayscaleValues[y * image.Width + x] = image[x, y].R / 255.0;
             }
         }
